Move e2e tests + harness under devinfra/js/debundle/e2e/
Single dir for the black-box e2e suite. Targets:

- //devinfra/js/debundle/e2e:lowering_test
- //devinfra/js/debundle/e2e:cross_module_test
- //devinfra/js/debundle/e2e:naturalization_test
- //devinfra/js/debundle/e2e:url_rebase_test

Drops the redundant `debundle_*_e2e_test` filename prefix and pulls
black_box.mjs out of test_support/ (it had no other consumer).
support.mjs and black_box.mjs sit next to the tests; their own
BUILD.bazel keeps the test data deps local.

BAZEL_TEST_INVOCATIONS=buildbuddy:fb4be489-8121-4033-984e-5820080a71d3
