+++ devinfra/js/debundle/e2e/black_box.mjs
